+++ src/Trarizon.Library/Mathematics/Geometry2D/Ray2D.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Numerics;
 
-namespace Trarizon.Library.Numerics;
+namespace Trarizon.Library.Mathematics.Geometry2D;
 public struct Ray2D(Vector2 origin, Vector2 direction) : IEquatable<Ray2D>
 #if NET7_0_OR_GREATER
     , IEqualityOperators<Ray2D, Ray2D, bool>