+++ src/Trarizon.Library/Mathematics/TraGeometry.2D.cs
@@ -1,0 +1,9 @@
+﻿using System.Numerics;
+#if NET7_0_OR_GREATER
+using Trarizon.Library.Mathematics.Geometry2D;
+#endif
+
+namespace Trarizon.Library.Mathematics;
+public static partial class TraGeometry
+{
+}