+++ src/Trarizon.Library/Mathematics/TraMath.Netstd.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Trarizon.Library.Mathematics;
+public static partial class TraMath
+{
+}