+++ src/Trarizon.Library/Mathematics/TraNumber.Mirror16.cs
@@ -4,10 +4,10 @@
 
 using CommunityToolkit.HighPerformance;
 
-namespace Trarizon.Library.Numerics;
+namespace Trarizon.Library.Mathematics;
 public static partial class TraNumber
 {
-    public static bool IncAndTryWrap(this ref int number, int delta, int max)
+    public static bool IncAndTryWrap(this ref short number, short delta, short max)
     {
         number += delta;
         if (number > max) {
@@ -19,7 +19,7 @@ public static partial class TraNumber
         }
     }
 
-    public static bool IncAndTryWrap(this ref uint number, uint delta, uint max)
+    public static bool IncAndTryWrap(this ref ushort number, ushort delta, ushort max)
     {
         number += delta;
         if (number > max) {
@@ -31,7 +31,7 @@ public static partial class TraNumber
         }
     }
 
-    public static void IncAndWrap(this ref int number, int delta, int max)
+    public static void IncAndWrap(this ref short number, short delta, short max)
     {
         number += delta;
         if (number > max) {
@@ -39,7 +39,7 @@ public static partial class TraNumber
         }
     }
 
-    public static void IncAndWrap(this ref uint number, uint delta, uint max)
+    public static void IncAndWrap(this ref ushort number, ushort delta, ushort max)
     {
         number += delta;
         if (number > max) {
@@ -52,18 +52,18 @@ public static partial class TraNumber
     /// </summary>
     /// <typeparam name="T"></typeparam>
     /// <param name="value"></param>
-    public static void FlipNegative(ref int value)
+    public static void FlipNegative(ref short value)
     {
         if (value < 0)
-            value = ~value;
+            value = (short)~value;
     }
 
     #region MinMax
 
-    public static int Min(int v0, int v1, int v2)
+    public static short Min(short v0, short v1, short v2)
         => v0 > v1 ? v1 : v0 > v2 ? v2 : v0;
 
-    public static int Min(params ReadOnlySpan<int> values) 
+    public static short Min(params ReadOnlySpan<short> values) 
     {
         var rtn = values[0];
         for (var i = 1; i < values.Length; i++) {
@@ -74,10 +74,10 @@ public static partial class TraNumber
         return rtn;
     }
 
-    public static uint Min(uint v0, uint v1, uint v2)
+    public static ushort Min(ushort v0, ushort v1, ushort v2)
         => v0 > v1 ? v1 : v0 > v2 ? v2 : v0;
 
-    public static uint Min(params ReadOnlySpan<uint> values) 
+    public static ushort Min(params ReadOnlySpan<ushort> values) 
     {
         var rtn = values[0];
         for (var i = 1; i < values.Length; i++) {
@@ -88,10 +88,10 @@ public static partial class TraNumber
         return rtn;
     }
 
-    public static int Max(int v0, int v1, int v2)
+    public static short Max(short v0, short v1, short v2)
         => v0 < v1 ? v1 : v0 < v2 ? v2 : v0;
 
-    public static int Max(params ReadOnlySpan<int> values) 
+    public static short Max(params ReadOnlySpan<short> values) 
     {
         var rtn = values[0];
         for (var i = 1; i < values.Length; i++) {
@@ -102,10 +102,10 @@ public static partial class TraNumber
         return rtn;
     }
 
-    public static uint Max(uint v0, uint v1, uint v2)
+    public static ushort Max(ushort v0, ushort v1, ushort v2)
         => v0 < v1 ? v1 : v0 < v2 ? v2 : v0;
 
-    public static uint Max(params ReadOnlySpan<uint> values) 
+    public static ushort Max(params ReadOnlySpan<ushort> values) 
     {
         var rtn = values[0];
         for (var i = 1; i < values.Length; i++) {
@@ -123,7 +123,7 @@ public static partial class TraNumber
     /// If <paramref name="left"/> equals <paramref name="right"/>, the return value is (<paramref name="left"/>, <paramref name="right"/>),
     /// else Min is the less one
     /// </returns>
-    public static (int Min, int Max) MinMax(int left, int right)
+    public static (short Min, short Max) MinMax(short left, short right)
     {
         if (left <= right)
             return (left, right);
@@ -131,7 +131,7 @@ public static partial class TraNumber
             return (right, left);
     }
 
-    public static (int Min, int Max) MinMax(params ReadOnlySpan<int> values) 
+    public static (short Min, short Max) MinMax(params ReadOnlySpan<short> values) 
     {
         var min = values[0];
         var max = min;
@@ -152,7 +152,7 @@ public static partial class TraNumber
     /// If <paramref name="left"/> equals <paramref name="right"/>, the return value is (<paramref name="left"/>, <paramref name="right"/>),
     /// else Min is the less one
     /// </returns>
-    public static (uint Min, uint Max) MinMax(uint left, uint right)
+    public static (ushort Min, ushort Max) MinMax(ushort left, ushort right)
     {
         if (left <= right)
             return (left, right);
@@ -160,7 +160,7 @@ public static partial class TraNumber
             return (right, left);
     }
 
-    public static (uint Min, uint Max) MinMax(params ReadOnlySpan<uint> values) 
+    public static (ushort Min, ushort Max) MinMax(params ReadOnlySpan<ushort> values) 
     {
         var min = values[0];
         var max = min;