+++ src/Trarizon.Library/Mathematics/TraNumber.MirrorFloat.cs
@@ -4,7 +4,7 @@
 
 using CommunityToolkit.HighPerformance;
 
-namespace Trarizon.Library.Numerics;
+namespace Trarizon.Library.Mathematics;
 public static partial class TraNumber
 {
     /// <summary>